coq_to_lean/pyproject.toml
[project]
name = "coq_to_lean"
version = "0.1.0"
description = "Default template for PDM package"
authors = [{ name = "Fady Adal", email = "2masadel@gmail.com" }]
dependencies = [
    "python-dotenv>=1.0.1",
    "lean-dojo>=1.5.1",
    "llama-cpp-python>=0.2.39",
]
readme = "README.md"
requires-python = "==3.10.*"
license = { text = "MIT" }

# [build-system]
# requires = ["hatchling"]
# build-backend = "hatchling.build"

[project.scripts]
main = "coq_to_lean:main"

[tool.pdm.scripts]
main = { call = "coq_to_lean:main" }

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["coq_to_lean"]
